# Conflicts:
#	Concussion Ball/Assets/Scripts/Camera/GUI/GUIHostMenu.cs
#	Concussion Ball/Assets/Scripts/Camera/GUI/GUIMainMenu.cs

+++ thomas/ThomasCore/src/thomas/WindowManager.cpp
@@ -97,7 +97,7 @@ namespace thomas
 
 	int WindowManager::GetNumOfWindows()
 	{
-		return m_windows.size();
+		return (int)m_windows.size();
 	}
 
 	Window* WindowManager::GetCurrentBound()
